tasks/requirements.txt
# ============================================================================
# INVOKE PYTHON PACKAGE REQUIREMENTS: For tasks
# ============================================================================
# DESCRIPTION:
#    pip install -r <THIS_FILE>
#
# SEE ALSO:
#  * http://www.pip-installer.org/
# ============================================================================

invoke >= 0.13.0
six >= 1.10.0
path.py >= 8.2.1
pathlib;    python_version <= '3.4'

# For cleanup of python files: py.cleanup
pycmd
